move auction.js and agencyEscrow.js out of core.
